refactor(admin-ui): Phase2-P1-2 A6 — KnowledgeListTab をライブ版で component 抽出 (機能変更なし) (GID 1214120401966111)
