+++ admin-ui/src/components/knowledge/KnowledgeListTab.tsx
@@ -3,20 +3,23 @@ import { useNavigate } from "react-router-dom";
 import KnowledgeFaqEditModal, { type KnowledgeFaqItem } from "../KnowledgeFaqEditModal";
 import { useLang } from "../../i18n/LangContext";
 import { API_BASE } from "../../lib/api";
-import {
-  type KnowledgeItem,
-  type DeleteState,
-  fetchWithAuth,
-  formatDate,
-  CARD_STYLE,
-  BTN_DANGER,
-} from "./shared";
-import FaqSearchBar from "./FaqSearchBar";
-import Pagination from "./Pagination";
-import BulkActionBar from "./BulkActionBar";
-import ExcludeSearchToggle from "./ExcludeSearchToggle";
+import { fetchWithAuth, formatDate, CARD_STYLE, BTN_DANGER } from "./shared";
 
-type SortKey = "created_at" | "updated_at" | "category";
+interface KnowledgeItem {
+  id: number;
+  tenant_id: string;
+  question: string;
+  answer: string;
+  category: string | null;
+  tags: string[] | null;
+  is_published?: boolean;
+  is_global?: boolean;
+  created_at: string;
+}
+
+type DeleteState = "idle" | "confirming" | "deleting" | "success" | "error";
+
+// ─── タブ1: ナレッジ一覧 ────────────────────────────────────────────────────
 
 export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
   const navigate = useNavigate();
@@ -30,26 +33,13 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
     { value: "store_info", label: t("category.store_info") },
   ];
 
-  // ─── List state ───────────────────────────────────────────────────────────
   const [items, setItems] = useState<KnowledgeItem[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
-  const [totalCount, setTotalCount] = useState(0);
-
-  // ─── Filter / sort / pagination state ────────────────────────────────────
   const [categoryFilter, setCategoryFilter] = useState<string>("all");
-  const [searchQuery, setSearchQuery] = useState("");
-  const [debouncedSearch, setDebouncedSearch] = useState("");
-  const [sortKey, setSortKey] = useState<SortKey>("created_at");
-  const [sortOrder, setSortOrder] = useState<"asc" | "desc">("desc");
-  const [currentOffset, setCurrentOffset] = useState(0);
-  const [pageLimit, setPageLimit] = useState(50);
-
-  // ─── Selection / bulk state ───────────────────────────────────────────────
-  const [selectedIds, setSelectedIds] = useState<Set<number>>(new Set());
-  const [isBulkDeleting, setIsBulkDeleting] = useState(false);
-
-  // ─── Modal / toast state ──────────────────────────────────────────────────
+  const [publishFilter, setPublishFilter] = useState<"all" | "published" | "draft">("all");
+  const [globalFilter, setGlobalFilter] = useState<"all" | "global" | "tenant">("all");
+  const [togglingId, setTogglingId] = useState<number | null>(null);
   const [deleteTarget, setDeleteTarget] = useState<{
     id: number;
     question: string;
@@ -60,70 +50,6 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
   const [createMode, setCreateMode] = useState(false);
   const [toast, setToast] = useState<string | null>(null);
 
-  // ─── Debounce search ──────────────────────────────────────────────────────
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebouncedSearch(searchQuery);
-      setCurrentOffset(0);
-      setSelectedIds(new Set());
-    }, 300);
-    return () => clearTimeout(timer);
-  }, [searchQuery]);
-
-  // ─── Fetch FAQs ───────────────────────────────────────────────────────────
-  const fetchFaqs = useCallback(async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const params = new URLSearchParams({
-        tenant: tenantId,
-        limit: String(pageLimit),
-        offset: String(currentOffset),
-        sort: sortKey,
-        order: sortOrder,
-      });
-      if (debouncedSearch) params.set("search", debouncedSearch);
-      if (categoryFilter !== "all") params.set("category", categoryFilter);
-
-      const res = await fetchWithAuth(
-        `${API_BASE}/v1/admin/knowledge/faq?${params}`
-      );
-      if (!res.ok) throw new Error(t("knowledge.load_error"));
-
-      const data = (await res.json()) as {
-        items?: KnowledgeItem[];
-        faqs?: KnowledgeItem[];
-        total: number;
-      };
-      setItems(data.faqs ?? data.items ?? []);
-      setTotalCount(data.total ?? 0);
-    } catch (err) {
-      if (err instanceof Error && err.message === "__AUTH_REQUIRED__") {
-        navigate("/login", { replace: true });
-        return;
-      }
-      setError(
-        err instanceof Error ? err.message : t("knowledge.load_error")
-      );
-    } finally {
-      setLoading(false);
-    }
-  }, [
-    navigate,
-    categoryFilter,
-    debouncedSearch,
-    sortKey,
-    sortOrder,
-    currentOffset,
-    pageLimit,
-    t,
-  ]);
-
-  useEffect(() => {
-    void fetchFaqs();
-  }, [fetchFaqs]);
-
-  // ─── Helpers ──────────────────────────────────────────────────────────────
   const showToast = (msg: string) => {
     setToast(msg);
     setTimeout(() => setToast(null), 3000);
@@ -133,43 +59,49 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
     setEditTarget(null);
     setCreateMode(false);
     showToast(msg);
-    void fetchFaqs();
+    void fetchItems();
   };
 
-  const categoryLabel = (cat: string | null) => {
-    const found = CATEGORIES.find((c) => c.value === cat);
-    return found ? found.label : cat ?? t("knowledge.uncategorized");
-  };
+  const fetchItems = useCallback(async () => {
+    setLoading(true);
+    setError(null);
+    try {
+      const params = new URLSearchParams({ tenant: tenantId });
+      if (categoryFilter !== "all") params.set("category", categoryFilter);
+      if (publishFilter === "published") params.set("is_published", "true");
+      if (publishFilter === "draft") params.set("is_published", "false");
+      if (globalFilter === "global") params.set("is_global", "true");
+      if (globalFilter === "tenant") params.set("is_global", "false");
 
-  // ─── Category / sort handlers ─────────────────────────────────────────────
-  const handleCategoryChange = (cat: string) => {
-    setCategoryFilter(cat);
-    setCurrentOffset(0);
-    setSelectedIds(new Set());
-  };
+      const res = await fetchWithAuth(`${API_BASE}/v1/admin/knowledge?${params}`);
+      if (!res.ok) throw new Error(t("knowledge.load_error"));
+      const data = (await res.json()) as { items: KnowledgeItem[] };
+      setItems(data.items ?? []);
+    } catch (err) {
+      if (err instanceof Error && err.message === "__AUTH_REQUIRED__") {
+        navigate("/login", { replace: true });
+        return;
+      }
+      setError(err instanceof Error ? err.message : t("knowledge.load_error"));
+    } finally {
+      setLoading(false);
+    }
+  }, [navigate, tenantId, categoryFilter, publishFilter, globalFilter, t]);
 
-  const handleSortChange = (value: string) => {
-    const parts = value.split("_");
-    const ord = parts.pop() as "asc" | "desc";
-    const key = parts.join("_") as SortKey;
-    setSortKey(key);
-    setSortOrder(ord);
-    setCurrentOffset(0);
-  };
+  useEffect(() => { fetchItems(); }, [fetchItems]);
 
-  // ─── Single-item delete ───────────────────────────────────────────────────
   const handleDelete = async () => {
     if (!deleteTarget) return;
-    setDeleteTarget((prev) => (prev ? { ...prev, state: "deleting" } : null));
+
+    setDeleteTarget((prev) => prev ? { ...prev, state: "deleting" } : null);
     try {
       const res = await fetchWithAuth(
         `${API_BASE}/v1/admin/knowledge/${deleteTarget.id}?tenant=${tenantId}`,
         { method: "DELETE" }
       );
       if (!res.ok) throw new Error(t("knowledge.delete_error"));
-      setDeleteTarget((prev) => (prev ? { ...prev, state: "success" } : null));
+      setDeleteTarget((prev) => prev ? { ...prev, state: "success" } : null);
       setItems((prev) => prev.filter((i) => i.id !== deleteTarget.id));
-      setTotalCount((prev) => Math.max(0, prev - 1));
       setTimeout(() => setDeleteTarget(null), 2000);
     } catch (err) {
       if (err instanceof Error && err.message === "__AUTH_REQUIRED__") {
@@ -177,108 +109,46 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
         return;
       }
       setDeleteTarget((prev) =>
-        prev
-          ? {
-              ...prev,
-              state: "error",
-              error:
-                err instanceof Error ? err.message : t("knowledge.delete_error"),
-            }
-          : null
+        prev ? { ...prev, state: "error", error: err instanceof Error ? err.message : t("knowledge.delete_error") } : null
       );
     }
   };
 
-  // ─── Selection helpers ────────────────────────────────────────────────────
-  const toggleSelect = (id: number) => {
-    setSelectedIds((prev) => {
-      const next = new Set(prev);
-      if (next.has(id)) {
-        next.delete(id);
-      } else {
-        next.add(id);
-      }
-      return next;
-    });
-  };
-
-  const toggleSelectAll = () => {
-    const pageIds = items.map((f) => f.id);
-    const allSelected =
-      pageIds.length > 0 && pageIds.every((id) => selectedIds.has(id));
-    setSelectedIds((prev) => {
-      const next = new Set(prev);
-      if (allSelected) {
-        pageIds.forEach((id) => next.delete(id));
-      } else {
-        pageIds.forEach((id) => next.add(id));
-      }
-      return next;
-    });
-  };
-
-  // ─── Bulk delete ──────────────────────────────────────────────────────────
-  const handleBulkDelete = async () => {
-    const count = selectedIds.size;
-    const msg =
-      lang === "ja"
-        ? `本当に${count}件のFAQを削除しますか？この操作は元に戻せません。`
-        : `Are you sure you want to delete ${count} FAQs? This cannot be undone.`;
-    if (!window.confirm(msg)) return;
-
-    setIsBulkDeleting(true);
+  const handleTogglePublish = async (item: KnowledgeItem) => {
+    setTogglingId(item.id);
     try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v1/admin/knowledge/faq/bulk?tenant=${tenantId}`,
+      const newState = !item.is_published;
+      await fetchWithAuth(
+        `${API_BASE}/v1/admin/knowledge/faq/${item.id}?tenant=${tenantId}`,
         {
-          method: "DELETE",
+          method: "PUT",
           headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ ids: Array.from(selectedIds) }),
+          body: JSON.stringify({
+            question: item.question,
+            answer: item.answer,
+            category: item.category ?? undefined,
+            tags: item.tags ?? [],
+            is_published: newState,
+          }),
         }
       );
-      if (res.ok) {
-        setSelectedIds(new Set());
-        showToast(
-          lang === "ja"
-            ? `${count}件のFAQを削除しました`
-            : `Deleted ${count} FAQs`
-        );
-        void fetchFaqs();
-      } else {
-        showToast(
-          lang === "ja"
-            ? "削除に失敗しました。もう一度お試しください"
-            : "Failed to delete. Please try again."
-        );
-      }
-    } catch {
-      showToast(
-        lang === "ja"
-          ? "削除に失敗しました。もう一度お試しください"
-          : "Failed to delete. Please try again."
+      setItems((prev) =>
+        prev.map((i) => (i.id === item.id ? { ...i, is_published: newState } : i))
       );
+    } catch {
+      // no-op
     } finally {
-      setIsBulkDeleting(false);
+      setTogglingId(null);
     }
   };
 
-  // ─── 検索除外フラグ楽観的更新 ────────────────────────────────────────────────
-  const handleExcludeToggled = (faqId: number, newValue: boolean) => {
-    setItems((prev) =>
-      prev.map((item) =>
-        item.id === faqId ? { ...item, is_excluded_from_search: newValue } : item
-      )
-    );
+  const categoryLabel = (cat: string | null) => {
+    const found = CATEGORIES.find((c) => c.value === cat);
+    return found ? found.label : cat ?? t("knowledge.uncategorized");
   };
 
-  // ─── Derived ──────────────────────────────────────────────────────────────
-  const pageIds = items.map((f) => f.id);
-  const allPageSelected =
-    pageIds.length > 0 && pageIds.every((id) => selectedIds.has(id));
-  const somePageSelected = pageIds.some((id) => selectedIds.has(id));
-
   return (
-    <div style={{ paddingBottom: selectedIds.size > 0 ? 88 : 0 }}>
+    <div>
       {/* 新規追加ボタン */}
       <button
         onClick={() => setCreateMode(true)}
@@ -288,8 +158,7 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
           minHeight: 60,
           borderRadius: 14,
           border: "none",
-          background:
-            "linear-gradient(135deg, #22c55e 0%, #4ade80 50%, #22c55e 100%)",
+          background: "linear-gradient(135deg, #22c55e 0%, #4ade80 50%, #22c55e 100%)",
           color: "#022c22",
           fontSize: 18,
           fontWeight: 700,
@@ -306,72 +175,29 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
         {t("knowledge.add_faq")}
       </button>
 
-      {/* 検索バー */}
-      <FaqSearchBar value={searchQuery} onChange={setSearchQuery} />
-
-      {/* カテゴリフィルター + ソート + 更新ボタン */}
-      <div
-        style={{
-          display: "flex",
-          gap: 8,
-          marginBottom: 16,
-          flexWrap: "wrap",
-          alignItems: "center",
-        }}
-      >
-        <span style={{ fontSize: 14, color: "#9ca3af" }}>
-          {t("knowledge.category_filter")}
-        </span>
-        {[{ value: "all", label: t("knowledge.all") }, ...CATEGORIES].map(
-          (c) => (
-            <button
-              key={c.value}
-              onClick={() => handleCategoryChange(c.value)}
-              style={{
-                padding: "6px 14px",
-                minHeight: 36,
-                borderRadius: 999,
-                border: `1px solid ${
-                  categoryFilter === c.value ? "#22c55e" : "#374151"
-                }`,
-                background:
-                  categoryFilter === c.value
-                    ? "rgba(34,197,94,0.15)"
-                    : "transparent",
-                color:
-                  categoryFilter === c.value ? "#4ade80" : "#9ca3af",
-                fontSize: 13,
-                cursor: "pointer",
-              }}
-            >
-              {c.label}
-            </button>
-          )
-        )}
-
-        {/* ソートドロップダウン */}
-        <select
-          value={`${sortKey}_${sortOrder}`}
-          onChange={(e) => handleSortChange(e.target.value)}
-          style={{
-            padding: "6px 10px",
-            minHeight: 36,
-            borderRadius: 8,
-            border: "1px solid #374151",
-            background: "rgba(15,23,42,0.8)",
-            color: "#9ca3af",
-            fontSize: 13,
-            cursor: "pointer",
-          }}
-        >
-          <option value="created_at_desc">{t("knowledge.sort_newest")}</option>
-          <option value="created_at_asc">{t("knowledge.sort_oldest")}</option>
-          <option value="updated_at_desc">{t("knowledge.sort_updated")}</option>
-          <option value="category_asc">{t("knowledge.sort_category")}</option>
-        </select>
-
+      {/* フィルター */}
+      <div style={{ display: "flex", gap: 8, marginBottom: 12, flexWrap: "wrap", alignItems: "center" }}>
+        <span style={{ fontSize: 14, color: "#9ca3af" }}>{t("knowledge.category_filter")}</span>
+        {[{ value: "all", label: t("knowledge.all") }, ...CATEGORIES].map((c) => (
+          <button
+            key={c.value}
+            onClick={() => setCategoryFilter(c.value)}
+            style={{
+              padding: "6px 14px",
+              minHeight: 36,
+              borderRadius: 999,
+              border: `1px solid ${categoryFilter === c.value ? "#22c55e" : "#374151"}`,
+              background: categoryFilter === c.value ? "rgba(34,197,94,0.15)" : "transparent",
+              color: categoryFilter === c.value ? "#4ade80" : "#9ca3af",
+              fontSize: 13,
+              cursor: "pointer",
+            }}
+          >
+            {c.label}
+          </button>
+        ))}
         <button
-          onClick={() => void fetchFaqs()}
+          onClick={fetchItems}
           disabled={loading}
           style={{
             marginLeft: "auto",
@@ -388,226 +214,165 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
           {loading ? t("knowledge.refreshing") : t("common.refresh")}
         </button>
       </div>
+      {/* 公開状態フィルター */}
+      <div style={{ display: "flex", gap: 6, marginBottom: 8, alignItems: "center" }}>
+        {(["all", "published", "draft"] as const).map((v) => {
+          const label = v === "all" ? (lang === "en" ? "All" : "すべて") : v === "published" ? (lang === "en" ? "Published" : "公開中") : (lang === "en" ? "Draft" : "非公開");
+          const active = publishFilter === v;
+          return (
+            <button
+              key={v}
+              onClick={() => setPublishFilter(v)}
+              style={{
+                padding: "4px 12px",
+                minHeight: 32,
+                borderRadius: 999,
+                border: `1px solid ${active ? "#3b82f6" : "#374151"}`,
+                background: active ? "rgba(59,130,246,0.15)" : "transparent",
+                color: active ? "#93c5fd" : "#6b7280",
+                fontSize: 12,
+                cursor: "pointer",
+              }}
+            >
+              {label}
+            </button>
+          );
+        })}
+      </div>
+      {/* グローバルナレッジフィルター */}
+      <div style={{ display: "flex", gap: 6, marginBottom: 16, alignItems: "center" }}>
+        {([
+          { v: "all", label: "全て" },
+          { v: "tenant", label: "テナント別" },
+          { v: "global", label: "⭐ グローバルのみ" },
+        ] as const).map(({ v, label }) => {
+          const active = globalFilter === v;
+          return (
+            <button
+              key={v}
+              onClick={() => setGlobalFilter(v)}
+              style={{
+                padding: "4px 12px",
+                minHeight: 32,
+                borderRadius: 999,
+                border: `1px solid ${active ? "rgba(234,179,8,0.5)" : "#374151"}`,
+                background: active ? "rgba(234,179,8,0.1)" : "transparent",
+                color: active ? "#fbbf24" : "#6b7280",
+                fontSize: 12,
+                cursor: "pointer",
+              }}
+            >
+              {label}
+            </button>
+          );
+        })}
+      </div>
 
       {error && (
-        <div
-          style={{
-            marginBottom: 16,
-            padding: "14px 18px",
-            borderRadius: 12,
-            background: "rgba(127,29,29,0.4)",
-            border: "1px solid rgba(248,113,113,0.3)",
-            color: "#fca5a5",
-            fontSize: 15,
-          }}
-        >
+        <div style={{ marginBottom: 16, padding: "14px 18px", borderRadius: 12, background: "rgba(127,29,29,0.4)", border: "1px solid rgba(248,113,113,0.3)", color: "#fca5a5", fontSize: 15 }}>
           {error}
         </div>
       )}
 
       {loading && items.length === 0 ? (
-        <div
-          style={{ padding: 40, textAlign: "center", color: "#6b7280" }}
-        >
-          <span style={{ display: "block", fontSize: 32, marginBottom: 8 }}>
-            ⏳
-          </span>
+        <div style={{ padding: 40, textAlign: "center", color: "#6b7280" }}>
+          <span style={{ display: "block", fontSize: 32, marginBottom: 8 }}>⏳</span>
           {t("knowledge.loading")}
         </div>
       ) : items.length === 0 ? (
-        <div
-          style={{
-            padding: 40,
-            textAlign: "center",
-            borderRadius: 14,
-            border: "1px dashed #374151",
-            background: "rgba(15,23,42,0.4)",
-          }}
-        >
-          <span style={{ display: "block", fontSize: 40, marginBottom: 12 }}>
-            📭
-          </span>
-          <p
-            style={{
-              fontSize: 16,
-              fontWeight: 600,
-              color: "#d1d5db",
-              margin: 0,
-            }}
-          >
+        <div style={{ padding: 40, textAlign: "center", borderRadius: 14, border: "1px dashed #374151", background: "rgba(15,23,42,0.4)" }}>
+          <span style={{ display: "block", fontSize: 40, marginBottom: 12 }}>📭</span>
+          <p style={{ fontSize: 16, fontWeight: 600, color: "#d1d5db", margin: 0 }}>
             {t("knowledge.empty_title")}
           </p>
-          <p
-            style={{
-              fontSize: 13,
-              color: "#6b7280",
-              marginTop: 6,
-              marginBottom: 0,
-            }}
-          >
+          <p style={{ fontSize: 13, color: "#6b7280", marginTop: 6, marginBottom: 0 }}>
             {t("knowledge.empty_sub")}
           </p>
         </div>
       ) : (
         <div style={{ ...CARD_STYLE, padding: 0, overflow: "hidden" }}>
-          {/* ヘッダー行: 全選択 + 件数 */}
-          <div
-            style={{
-              padding: "12px 18px",
-              borderBottom: "1px solid #111827",
-              fontSize: 13,
-              color: "#6b7280",
-              display: "flex",
-              alignItems: "center",
-              gap: 12,
-            }}
-          >
-            <label
-              style={{
-                display: "flex",
-                alignItems: "center",
-                gap: 8,
-                cursor: "pointer",
-                minHeight: 32,
-                userSelect: "none",
-              }}
-            >
-              <input
-                type="checkbox"
-                checked={allPageSelected}
-                ref={(el) => {
-                  if (el)
-                    el.indeterminate = somePageSelected && !allPageSelected;
-                }}
-                onChange={toggleSelectAll}
-                style={{ width: 18, height: 18, cursor: "pointer", accentColor: "#22c55e" }}
-              />
-              <span>{t("knowledge.select_all")}</span>
-            </label>
-            <span style={{ marginLeft: "auto" }}>
-              {t("knowledge.count", { n: totalCount })}
-            </span>
+          <div style={{ padding: "12px 18px", borderBottom: "1px solid #111827", fontSize: 13, color: "#6b7280" }}>
+            {t("knowledge.count", { n: items.length })}
           </div>
-
-          {/* FAQ行 */}
           {items.map((item, idx) => (
             <div
               key={item.id}
               style={{
                 padding: "16px 18px",
-                borderBottom:
-                  idx === items.length - 1 ? "none" : "1px solid #111827",
+                borderBottom: idx === items.length - 1 ? "none" : "1px solid #111827",
                 display: "flex",
-                gap: 12,
+                gap: 14,
                 alignItems: "flex-start",
                 flexWrap: "wrap",
-                background: selectedIds.has(item.id)
-                  ? "rgba(34,197,94,0.04)"
-                  : item.is_excluded_from_search
-                  ? "rgba(127,29,29,0.06)"
-                  : "transparent",
-                opacity: item.is_excluded_from_search ? 0.75 : 1,
-                transition: "all 0.15s",
+                opacity: item.is_published === false ? 0.55 : 1,
+                transition: "opacity 0.2s",
               }}
             >
-              {/* チェックボックス */}
-              <div style={{ paddingTop: 2, flexShrink: 0 }}>
-                <input
-                  type="checkbox"
-                  checked={selectedIds.has(item.id)}
-                  onChange={() => toggleSelect(item.id)}
-                  style={{
-                    width: 18,
-                    height: 18,
-                    cursor: "pointer",
-                    accentColor: "#22c55e",
-                  }}
-                />
-              </div>
-
               <div style={{ flex: 1, minWidth: 200 }}>
-                <div
-                  style={{
-                    display: "flex",
-                    gap: 8,
-                    alignItems: "center",
-                    flexWrap: "wrap",
-                    marginBottom: 6,
-                  }}
-                >
-                  <span
-                    style={{
-                      padding: "2px 8px",
-                      borderRadius: 999,
-                      background: "rgba(34,197,94,0.1)",
-                      border: "1px solid rgba(34,197,94,0.2)",
-                      color: "#4ade80",
-                      fontSize: 11,
-                      fontWeight: 600,
-                    }}
-                  >
+                <div style={{ display: "flex", gap: 8, alignItems: "center", flexWrap: "wrap", marginBottom: 6 }}>
+                  <span style={{
+                    padding: "2px 8px",
+                    borderRadius: 999,
+                    background: "rgba(34,197,94,0.1)",
+                    border: "1px solid rgba(34,197,94,0.2)",
+                    color: "#4ade80",
+                    fontSize: 11,
+                    fontWeight: 600,
+                  }}>
                     {categoryLabel(item.category)}
                   </span>
-                  {item.is_excluded_from_search && (
-                    <span
-                      style={{
-                        padding: "2px 8px",
-                        borderRadius: 999,
-                        background: "rgba(239,68,68,0.1)",
-                        border: "1px solid rgba(239,68,68,0.3)",
-                        color: "#f87171",
-                        fontSize: 11,
-                        fontWeight: 600,
-                      }}
-                    >
-                      検索除外
+                  <span style={{
+                    padding: "2px 8px",
+                    borderRadius: 999,
+                    background: item.is_published === false ? "rgba(75,85,99,0.3)" : "rgba(34,197,94,0.08)",
+                    border: `1px solid ${item.is_published === false ? "#4b5563" : "rgba(34,197,94,0.2)"}`,
+                    color: item.is_published === false ? "#6b7280" : "#86efac",
+                    fontSize: 11,
+                    fontWeight: 600,
+                  }}>
+                    {item.is_published === false
+                      ? (lang === "en" ? "⏸️ Draft" : "⏸️ 非公開")
+                      : (lang === "en" ? "✅ Published" : "✅ 公開中")}
+                  </span>
+                  {item.is_global && (
+                    <span style={{
+                      padding: "2px 8px", borderRadius: 999, fontSize: 11, fontWeight: 600,
+                      background: "rgba(234,179,8,0.12)", border: "1px solid rgba(234,179,8,0.3)", color: "#fbbf24",
+                    }}>
+                      ⭐ グローバル
                     </span>
                   )}
-                  <span style={{ fontSize: 11, color: "#6b7280" }}>
-                    {formatDate(item.created_at, locale)}
-                  </span>
+                  <span style={{ fontSize: 11, color: "#6b7280" }}>{formatDate(item.created_at, locale)}</span>
                 </div>
-                <p
-                  style={{
-                    fontSize: 15,
-                    fontWeight: 600,
-                    color: "#f9fafb",
-                    margin: "0 0 4px",
-                    lineHeight: 1.4,
-                  }}
-                >
+                <p style={{ fontSize: 15, fontWeight: 600, color: "#f9fafb", margin: "0 0 4px", lineHeight: 1.4 }}>
                   Q: {item.question}
                 </p>
-                <p
-                  style={{
-                    fontSize: 13,
-                    color: "#9ca3af",
-                    margin: 0,
-                    lineHeight: 1.5,
-                  }}
-                >
-                  A:{" "}
-                  {item.answer.slice(0, 120)}
-                  {item.answer.length > 120 ? "…" : ""}
+                <p style={{ fontSize: 13, color: "#9ca3af", margin: 0, lineHeight: 1.5 }}>
+                  A: {item.answer.slice(0, 120)}{item.answer.length > 120 ? "…" : ""}
                 </p>
               </div>
-
-              <div
-                style={{
-                  display: "flex",
-                  gap: 8,
-                  flexShrink: 0,
-                  alignItems: "center",
-                  flexWrap: "wrap",
-                }}
-              >
-                <ExcludeSearchToggle
-                  faqId={item.id}
-                  tenantId={tenantId}
-                  isExcluded={item.is_excluded_from_search ?? false}
-                  onToggled={handleExcludeToggled}
-                  onError={(msg) => showToast(msg)}
-                />
+              <div style={{ display: "flex", gap: 8, flexShrink: 0, alignItems: "center", flexWrap: "wrap" }}>
+                <button
+                  onClick={() => handleTogglePublish(item)}
+                  disabled={togglingId === item.id}
+                  style={{
+                    padding: "10px 14px",
+                    minHeight: 44,
+                    borderRadius: 10,
+                    border: `1px solid ${item.is_published === false ? "rgba(34,197,94,0.4)" : "#4b5563"}`,
+                    background: item.is_published === false ? "rgba(34,197,94,0.1)" : "rgba(75,85,99,0.15)",
+                    color: item.is_published === false ? "#4ade80" : "#9ca3af",
+                    fontSize: 13,
+                    cursor: togglingId === item.id ? "default" : "pointer",
+                    fontWeight: 500,
+                    whiteSpace: "nowrap",
+                    opacity: togglingId === item.id ? 0.6 : 1,
+                  }}
+                >
+                  {item.is_published === false
+                    ? (lang === "en" ? "Publish" : "公開する")
+                    : (lang === "en" ? "Unpublish" : "非公開にする")}
+                </button>
                 <button
                   onClick={() =>
                     setEditTarget({
@@ -617,7 +382,6 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
                       category: item.category,
                       tags: item.tags,
                       is_published: item.is_published,
-                      is_excluded_from_search: item.is_excluded_from_search,
                     })
                   }
                   style={{
@@ -636,13 +400,7 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
                   {t("knowledge.edit")}
                 </button>
                 <button
-                  onClick={() =>
-                    setDeleteTarget({
-                      id: item.id,
-                      question: item.question,
-                      state: "confirming",
-                    })
-                  }
+                  onClick={() => setDeleteTarget({ id: item.id, question: item.question, state: "confirming" })}
                   style={BTN_DANGER}
                 >
                   {t("knowledge.delete")}
@@ -650,34 +408,9 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
               </div>
             </div>
           ))}
-
-          {/* ページネーション */}
-          <div style={{ padding: "0 18px" }}>
-            <Pagination
-              total={totalCount}
-              limit={pageLimit}
-              offset={currentOffset}
-              onPageChange={(newOffset) => {
-                setCurrentOffset(newOffset);
-              }}
-              onLimitChange={(newLimit) => {
-                setPageLimit(newLimit);
-                setCurrentOffset(0);
-              }}
-            />
-          </div>
         </div>
       )}
 
-      {/* バルク操作バー */}
-      <BulkActionBar
-        selectedCount={selectedIds.size}
-        onBulkDelete={handleBulkDelete}
-        onClearSelection={() => setSelectedIds(new Set())}
-        loading={isBulkDeleting}
-      />
-
-      {/* 編集モーダル */}
       {editTarget && (
         <KnowledgeFaqEditModal
           mode="edit"
@@ -688,7 +421,6 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
         />
       )}
 
-      {/* 新規作成モーダル */}
       {createMode && (
         <KnowledgeFaqEditModal
           mode="create"
@@ -698,12 +430,11 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
         />
       )}
 
-      {/* トースト通知 */}
       {toast && (
         <div
           style={{
             position: "fixed",
-            bottom: selectedIds.size > 0 ? 104 : 32,
+            bottom: 32,
             left: "50%",
             transform: "translateX(-50%)",
             zIndex: 2000,
@@ -722,102 +453,26 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
         </div>
       )}
 
-      {/* 削除確認ダイアログ */}
       {deleteTarget && (
         <div
-          style={{
-            position: "fixed",
-            inset: 0,
-            background: "rgba(0,0,0,0.7)",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            zIndex: 1000,
-            padding: 20,
-          }}
-          onClick={(e) => {
-            if (
-              e.target === e.currentTarget &&
-              deleteTarget.state !== "deleting"
-            )
-              setDeleteTarget(null);
-          }}
+          style={{ position: "fixed", inset: 0, background: "rgba(0,0,0,0.7)", display: "flex", alignItems: "center", justifyContent: "center", zIndex: 1000, padding: 20 }}
+          onClick={(e) => { if (e.target === e.currentTarget && deleteTarget.state !== "deleting") setDeleteTarget(null); }}
         >
-          <div
-            style={{
-              background: "#0f172a",
-              border: "1px solid #1f2937",
-              borderRadius: 16,
-              padding: "28px 24px",
-              maxWidth: 420,
-              width: "100%",
-              boxShadow: "0 24px 60px rgba(0,0,0,0.6)",
-            }}
-          >
+          <div style={{ background: "#0f172a", border: "1px solid #1f2937", borderRadius: 16, padding: "28px 24px", maxWidth: 420, width: "100%", boxShadow: "0 24px 60px rgba(0,0,0,0.6)" }}>
             {deleteTarget.state === "success" ? (
               <div style={{ textAlign: "center" }}>
-                <span
-                  style={{
-                    fontSize: 48,
-                    display: "block",
-                    marginBottom: 12,
-                  }}
-                >
-                  ✅
-                </span>
-                <p
-                  style={{
-                    fontSize: 17,
-                    fontWeight: 600,
-                    color: "#4ade80",
-                    margin: 0,
-                  }}
-                >
-                  {t("knowledge.deleted")}
-                </p>
+                <span style={{ fontSize: 48, display: "block", marginBottom: 12 }}>✅</span>
+                <p style={{ fontSize: 17, fontWeight: 600, color: "#4ade80", margin: 0 }}>{t("knowledge.deleted")}</p>
               </div>
             ) : (
               <>
-                <h3
-                  style={{
-                    fontSize: 18,
-                    fontWeight: 700,
-                    color: "#f9fafb",
-                    margin: "0 0 12px",
-                  }}
-                >
-                  {t("knowledge.delete_confirm_title")}
-                </h3>
-                <p
-                  style={{
-                    fontSize: 14,
-                    color: "#d1d5db",
-                    margin: "0 0 6px",
-                  }}
-                >
-                  Q: {deleteTarget.question}
-                </p>
-                <p
-                  style={{
-                    fontSize: 13,
-                    color: "#9ca3af",
-                    margin: "0 0 20px",
-                    lineHeight: 1.6,
-                  }}
-                >
+                <h3 style={{ fontSize: 18, fontWeight: 700, color: "#f9fafb", margin: "0 0 12px" }}>{t("knowledge.delete_confirm_title")}</h3>
+                <p style={{ fontSize: 14, color: "#d1d5db", margin: "0 0 6px" }}>Q: {deleteTarget.question}</p>
+                <p style={{ fontSize: 13, color: "#9ca3af", margin: "0 0 20px", lineHeight: 1.6 }}>
                   {t("knowledge.delete_confirm_body")}
                 </p>
                 {deleteTarget.state === "error" && deleteTarget.error && (
-                  <div
-                    style={{
-                      marginBottom: 16,
-                      padding: "10px 14px",
-                      borderRadius: 8,
-                      background: "rgba(127,29,29,0.4)",
-                      color: "#fca5a5",
-                      fontSize: 14,
-                    }}
-                  >
+                  <div style={{ marginBottom: 16, padding: "10px 14px", borderRadius: 8, background: "rgba(127,29,29,0.4)", color: "#fca5a5", fontSize: 14 }}>
                     {deleteTarget.error}
                   </div>
                 )}
@@ -825,44 +480,16 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
                   <button
                     onClick={() => setDeleteTarget(null)}
                     disabled={deleteTarget.state === "deleting"}
-                    style={{
-                      flex: 1,
-                      padding: "14px",
-                      minHeight: 56,
-                      borderRadius: 10,
-                      border: "1px solid #374151",
-                      background: "transparent",
-                      color: "#e5e7eb",
-                      fontSize: 15,
-                      fontWeight: 600,
-                      cursor: "pointer",
-                    }}
+                    style={{ flex: 1, padding: "14px", minHeight: 56, borderRadius: 10, border: "1px solid #374151", background: "transparent", color: "#e5e7eb", fontSize: 15, fontWeight: 600, cursor: "pointer" }}
                   >
                     {t("knowledge.cancel_delete")}
                   </button>
                   <button
                     onClick={handleDelete}
                     disabled={deleteTarget.state === "deleting"}
-                    style={{
-                      flex: 1,
-                      padding: "14px",
-                      minHeight: 56,
-                      borderRadius: 10,
-                      border: "none",
-                      background:
-                        "linear-gradient(135deg, #991b1b, #dc2626)",
-                      color: "#fee2e2",
-                      fontSize: 15,
-                      fontWeight: 700,
-                      cursor:
-                        deleteTarget.state === "deleting"
-                          ? "not-allowed"
-                          : "pointer",
-                    }}
+                    style={{ flex: 1, padding: "14px", minHeight: 56, borderRadius: 10, border: "none", background: "linear-gradient(135deg, #991b1b, #dc2626)", color: "#fee2e2", fontSize: 15, fontWeight: 700, cursor: deleteTarget.state === "deleting" ? "not-allowed" : "pointer" }}
                   >
-                    {deleteTarget.state === "deleting"
-                      ? t("common.deleting")
-                      : t("knowledge.confirm_delete")}
+                    {deleteTarget.state === "deleting" ? t("common.deleting") : t("knowledge.confirm_delete")}
                   </button>
                 </div>
               </>
@@ -870,9 +497,6 @@ export default function KnowledgeListTab({ tenantId }: { tenantId: string }) {
           </div>
         </div>
       )}
-
-      {/* spin animation for ExcludeSearchToggle */}
-      <style>{`@keyframes spin { to { transform: rotate(360deg); } }`}</style>
     </div>
   );
 }
